agri_project-main/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-nested-routers==0.93.5
psycopg2-binary==2.9.10
openpyxl==3.1.5
pandas==2.2.3

